mkfs error handling

+++ src/standalonefat.c
@@ -1,3 +1,4 @@
+#include <errno.h>
 #include "pennfat.h"
 
 int main(int argc, char* argv[]) {
@@ -90,8 +91,28 @@ int main(int argc, char* argv[]) {
       // other test stuff
 
       else if (strcmp(args[0], "mkfs") == 0) {
+        if (args[1] == NULL || args[2] == NULL || args[3] == NULL) {
+          perror("error: invalid arguments for mkfs");
+          continue;
+        }
+        errno = 0;
         int blocks_in_fat = strtol(args[2], &ptr, base);
+        if (errno == ERANGE) {
+          perror("error: invalid arguments for number of blocks in fat");
+          continue;
+        }
+        errno = 0;
+        if (blocks_in_fat < 1 || blocks_in_fat > 32) {
+          perror("error: number of blocks in fat must be between 1~32");
+          continue;
+        }
         int block_size_config = strtol(args[3], &ptr, base);
+        if (errno == ERANGE) {
+          perror(
+              "error: invalid arguments for number of block size "
+              "configuration");
+          continue;
+        }
         mkfs(args[1], blocks_in_fat, block_size_config);
         // fprintf(stderr, "block size: %d\n", block_size);
       } else if (strcmp(args[0], "mount") == 0) {
@@ -108,26 +129,30 @@ int main(int argc, char* argv[]) {
         } else {
           mounted = false;
         }
-      } else if (strcmp(args[0], "write") == 0) {
-        int fd = strtol(args[1], &ptr, base);
-        k_write(fd, args[2], strlen(args[2]));
-      } else if (strcmp(args[0], "open") == 0) {
-        int mode = strtol(args[2], &ptr, base);
-        k_open(args[1], mode);
-      } else if (strcmp(args[0], "read") == 0) {
-        int fd = strtol(args[1], &ptr, base);
-        char buffer[1000];
-        buffer[999] = '\0';
-        k_read(fd, 1000, buffer);
-        fprintf(stderr, "READ OUTPUT: %s\n", buffer);
-      } else if (strcmp(args[0], "kunlink") == 0) {
-        k_unlink(args[1]);
-      } else if (strcmp(args[0], "klseek") == 0) {
-        int fd = strtol(args[1], &ptr, base);
-        int offset = strtol(args[2], &ptr, base);
-        int whence = strtol(args[3], &ptr, base);
-        k_lseek(fd, offset, whence);
-      } else {
+      }
+
+      // else if (strcmp(args[0], "write") == 0) {
+      //   int fd = strtol(args[1], &ptr, base);
+      //   k_write(fd, args[2], strlen(args[2]));
+      // } else if (strcmp(args[0], "open") == 0) {
+      //   int mode = strtol(args[2], &ptr, base);
+      //   k_open(args[1], mode);
+      // } else if (strcmp(args[0], "read") == 0) {
+      //   int fd = strtol(args[1], &ptr, base);
+      //   char buffer[1000];
+      //   buffer[999] = '\0';
+      //   k_read(fd, 1000, buffer);
+      //   fprintf(stderr, "READ OUTPUT: %s\n", buffer);
+      // } else if (strcmp(args[0], "kunlink") == 0) {
+      //   k_unlink(args[1]);
+      // } else if (strcmp(args[0], "klseek") == 0) {
+      //   int fd = strtol(args[1], &ptr, base);
+      //   int offset = strtol(args[2], &ptr, base);
+      //   int whence = strtol(args[3], &ptr, base);
+      //   k_lseek(fd, offset, whence);
+      // }
+
+      else {
         fprintf(stderr, "pennfat: command not found: %s\n", args[0]);
       }
       free_parsed_command(parsed);
